export_context argument made optional in exporters
