+++ doc/newsfragments/2528_new.export_context.rst
@@ -1,1 +1,1 @@
-Introduced new :py:class:`ExportContext <testplan.common.exporters.ExportContext>` class for exporters to communicate with each other.
+Introduced new :py:class:`ExportContext <testplan.common.exporters.ExportContext>` class for exporters to communicate with each other. Also exporters now return with an ExporterResult object. Custom exporters may need to be adjusted accordingly.
